WIP: Standalone Controllers/React Interface
