--- EcommerceApp/Controllers/HomeController.cs
@@ -1,71 +1,0 @@
-﻿using EcommerceApp.APIParsing;
-using EcommerceApp.Models;
-using EcommerceApp.Models.Interfaces;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Net.Http;
-using System.Threading.Tasks;
-
-namespace EcommerceApp.Controllers
-{
-  [Route("api/[controller]")]
-  [ApiController]
-  public class HomeController : ControllerBase
-  {
-    private readonly IGame _gameRepository;
-    private HttpClient httpClient = new HttpClient();
-    public HomeController(IGame gamerepo)
-    {
-      _gameRepository = gamerepo;
-    }
-
-    [HttpGet]
-    [AllowAnonymous]
-    public IActionResult RootRoute()
-    {
-      return Ok("Welcome to Ecommerce Games API");
-    }
-    
-    [HttpGet("allgames/")]
-    [AllowAnonymous]
-    public async Task<ActionResult<IEnumerable<Game>>> GetAllGames()
-    {
-      List<Game> getGames = await _gameRepository.GetAllGames();
-      return Ok(getGames);
-    }
-    [HttpGet("{id}")]
-    [AllowAnonymous]
-    public async Task<ActionResult<Game>> GetGame(int id)
-    {
-      Game game = await _gameRepository.GetGame(id);
-      return Ok(game);
-    }
-
-    [HttpGet("gamesdb/")]
-    [AllowAnonymous]
-    public async Task<ActionResult<String>> GetGameAPI()
-    {
-      string APIKEY = "";
-      string qString = "Super%20Metroid";
-      string URL = $"https://api.thegamesdb.net/v1/Games/ByGameName?apikey={APIKEY}&name={qString}"
-        + $"&fields=players%2Cpublishers%2Cgenres%2Coverview%2Clast_updated%2Crating%2Cplatform%2Ccoop%2Cyoutube%2Cos%2Cprocessor" +
-        "%2Cram%2Chee%2Cvideo%2Csound%2Calternates&include=boxart%2Cplatform";
-      Debug.WriteLine("URL: " + URL);
-
-      HttpResponseMessage response = await httpClient.GetAsync(URL);
-      string content = await response.Content.ReadAsStringAsync();
-
-      string returnData = $"{response.StatusCode} : {content}";
-      return Ok(returnData);
-    }
-    [HttpGet("testGamesDB/")]
-    [AllowAnonymous]
-    public async Task<ActionResult<string>> GetSampleGameAPI()
-    {
-      return Ok(TheGamesDBParse.SearchAPI("sample"));
-    }
-  }
-}
