+++ EcommerceApp/EcommerceApp.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="AuthorizeNet" Version="2.0.1" />
+    <PackageReference Include="AuthorizeNet" Version="2.0.2" />
     <PackageReference Include="Azure.Storage.Blobs" Version="12.8.0" />
     <PackageReference Include="Azure.Storage.Files.Shares" Version="12.1.0" />
     <PackageReference Include="Azure.Storage.Queues" Version="12.2.0" />
@@ -41,6 +41,10 @@
     <Folder Include="wwwroot\css\" />
     <Folder Include="wwwroot\images\" />
     <Folder Include="wwwroot\js\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\APIHelperFunctions\APIHelperFunctions.csproj" />
   </ItemGroup>
 
 </Project>
