+++ EcommerceApp/appsettings.json
@@ -6,10 +6,7 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=tcp:ecommerceappdbserver.database.windows.net,1433;Initial Catalog=EcommerceApp_db;Persist Security Info=False;User ID=admindatabase;Password=Temp1234;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
-
-  }
+  "AllowedHosts": "*"
+  
   
 }
